refactor(notification): move notification hooks to folder and adjust type/endpoint
