+++ FinEdgeFrontend/src/Hooks/Notifications/useGetNotifications.tsx
@@ -1,13 +1,12 @@
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import axios, { AxiosResponse, AxiosError } from "axios";
 import { HubConnectionBuilder } from "@microsoft/signalr";
-import { getNotificationHubEndPoint, getNotificationsEndPoint } from "../endpoints";
 import { useEffect } from "react";
-import { AppNotification } from "../Utils/Types";
-import useGetUser from "./useGetUser";
+import { getNotificationsEndPoint, getNotificationHubEndPoint } from "../../endpoints";
+import { AppNotification } from "../../Utils/Types";
 
 const getNotifications = async () => {
-return await axios
+  return await axios
     .get<AppNotification[]>(`${getNotificationsEndPoint}`, { withCredentials: true })
     .then((res: AxiosResponse<AppNotification[]>) => res.data)
     .catch((err: AxiosError) => {
@@ -16,14 +15,11 @@ return await axios
 };
 
 const useGetNotifications = () => {
-  const { data: user } = useGetUser();
-  
   const queryClient = useQueryClient();
 
   const notificationsQuery = useQuery({
     queryKey: ["notifications"],
     queryFn: getNotifications,
-    enabled: !!user,
   });
 
   const { data, isLoading, isError, error } = notificationsQuery;
