+++ FinEdgeFrontend/src/endpoints.ts
@@ -59,6 +59,8 @@ export const getPromptSuggestionsEndPoint = `${promptSuggestionController}/get`;
 // Notifications
 export const getNotificationsEndPoint = `${notificationController}/get-all`;
 export const getLatestUnreadNotificationEndPoint = `${notificationController}/get-latest-unread`;
-export const markNotificationAsReadAsync = `${notificationController}/mark-as-read`;
-export const markAllNotificationsAsReadAsync = `${notificationController}/mark-all-as-read`;
+export const markNotificationAsReadEndPoint = `${notificationController}/mark-as-read`;
+export const markAllNotificationsAsReadEndPoint = `${notificationController}/mark-all-as-read`;
+export const deleteNotificationEndPoint = `${notificationController}/delete`;
+export const deleteAllNotificationsEndPoint = `${notificationController}/delete-all`;
 export const getNotificationHubEndPoint = `${baseURL}/notificationHub`;
